cockroachdb/pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "plux>=1.3.1"]
build-backend = "setuptools.build_meta"

[project]
name = "localstack-extension-cockroachdb"
version = "0.1.0"
description = "LocalStack Extension: CockroachDB on LocalStack"
readme = {file = "README.md", content-type = "text/markdown; charset=UTF-8"}
requires-python = ">=3.10"
authors = [
    { name = "LocalStack team"}
]
keywords = ["LocalStack", "CockroachDB", "PostgreSQL", "SQL", "Distributed"]
classifiers = []
dependencies = [
    "localstack-extensions-utils"
]

[project.urls]
Homepage = "https://github.com/localstack/localstack-extensions"

[project.optional-dependencies]
dev = [
    "boto3",
    "build",
    "jsonpatch",
    "localstack-core",
    "psycopg2-binary",
    "pytest",
    "rolo",
    "ruff",
]

[project.entry-points."localstack.extensions"]
localstack_cockroachdb = "localstack_cockroachdb.extension:CockroachDbExtension"
